requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-silk==5.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gprof2dot==2024.6.6
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
